requirements/local.txt
# Local development dependencies go here
-r base.txt

django-debug-toolbar==1.4

pytest==2.9.1
pytest-django==2.9.1
coverage==4.0.3
prospector==0.11.7
